% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-status.R
\name{gh_api_status}
\alias{gh_api_status}
\title{GitHub APIs status}
\usage{
gh_api_status()
}
\value{
A character vector, one of "operational", "degraded_performance",
"partial_outage", or "major_outage."
}
\description{
Get the status of requests to GitHub APIs
}
\details{
See details in https://www.githubstatus.com/api#components.
}
\examples{
\dontrun{
gh_api_status()
}
}
